Refactor implement @Validated with specific groups depending on the function, and add @IdFormatConstraint to the parameters that need it

+++ Backend/UrbanStyle/src/main/java/io/github/juniorcorzo/UrbanStyle/domain/entities/ProductEntity.java
@@ -24,7 +24,7 @@ public class ProductEntity {
     @TextIndexed
     private String description;
     private double price;
-    private double discount;
+    private byte discount;
     private List<String> images;
     @TextIndexed
     private List<CategorySummary> categories;
